some more refactoring and added date to highscore
