+++ src/components/game/game-component.css
@@ -1,0 +1,9 @@
+.status {
+  margin-bottom: 10px;
+}
+
+.game {
+  display: flex;
+  flex-direction: column;
+}
+
