+++ src/components/highscores/highscores-component.js
@@ -1,16 +1,24 @@
 import './highscores-component.css'
 export function HighscoresComponent(props) {
+
+  
+
   return (
     <div className="highscores">
       <h2>Highscores</h2>
       <div className="centering">
       <table>
-        <thead><tr><th>Score</th><th>Name</th></tr></thead>
+        <thead><tr><th>Score</th><th>Name</th><th>Date</th></tr></thead>
         <tbody>
-        {props.scores.scores.map((score, ix) => {return(<tr className="score-row" key={ix}><td>{score.score}</td><td>{score.name}</td></tr>)})}
+        {props.scores.scores.map((score, ix) => {return(<tr className="score-row" key={ix}><td>{score.score}</td><td>{score.name}</td><td>{timeToString(score.time)}</td></tr>)})}
         </tbody>
       </table>
       </div>
     </div>
   )
+}
+
+function timeToString(scoreTime) {
+  const [day, month, date, year, time] = scoreTime.split(" ")
+  return `${date} ${month} ${year} - ${time}`
 }
